% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtData_doc.R
\name{gtData}
\alias{gtData}
\title{Group Testing Data Structure}
\description{
To fit group testing models in a versatile manner, we developed a data structure called \code{gtData} that contains pooled test outcomes and other relevant information about a group testing application. This data structure is used in several functions in this package, including \code{\link{glm.gt}} and \code{\link{prop.gt}}.
}
\details{
\code{gtData} must be specified as follows. Columns 1-5 consist of the pooled test outcomes (0 for negative and 1 for positive), pool sizes, pool-specific sensitivities, pool-specific specificities, and assay identification (ID) numbers, respectively. From column 6 onward, the pool member ID numbers must be specified.

\preformatted{
       Z psz   Se   Sp Assay Mem1 Mem2 Mem3 Mem4 Mem5 Mem6
Pool:1 1   6 0.90 0.92     1    1    2    3    4    5    6
Pool:2 0   6 0.90 0.92     1    7    8    9   10   11   12
Pool:3 1   2 0.95 0.96     2    1    2   -9   -9   -9   -9
Pool:4 0   2 0.95 0.96     2    3    4   -9   -9   -9   -9
Pool:5 1   2 0.95 0.96     2    5    6   -9   -9   -9   -9
Pool:6 0   1 0.92 0.90     3    1   -9   -9   -9   -9   -9
Pool:7 1   1 0.92 0.90     3    2   -9   -9   -9   -9   -9
Pool:8 0   1 0.92 0.90     3    5   -9   -9   -9   -9   -9
Pool:9 0   1 0.92 0.90     3    6   -9   -9   -9   -9   -9
}

Pool member IDs must start at 1 and increase consecutively up to \eqn{N},
the total number of individuals tested. For smaller pools, unused member ID
positions should be filled with \code{-9} or any other finite non-positive number as shown in the example above.
}
