% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtcode.R
\name{hier.gt.simulation}
\alias{hier.gt.simulation}
\title{Simulating Hierarchical Group Testing Data}
\usage{
hier.gt.simulation(N, p = 0.1, S, psz, Se, Sp, assayID, Yt = NULL)
}
\arguments{
\item{N}{The number of individuals to be tested.}

\item{p}{A scalar or a vector of length N consisting of individual disease probabilities.}

\item{S}{The number of stages used in testing, where \code{S} >= 1.}

\item{psz}{A vector of pool sizes in stages 1-\code{S}.}

\item{Se}{A vector of assay sensitivities in stages 1-\code{S}.}

\item{Sp}{A vector of assay specificities in stages 1-\code{S}.}

\item{assayID}{A vector of the identification numbers of the assays used in stages 1-\code{S}.}

\item{Yt}{A vector of individual true disease statuses.}
}
\value{
A list with components:
\item{gtData}{The simulated group testing data.}
\item{testsExp}{The number of tests expended.}
}
\description{
This function simulates hierarchical group testing data with any number of hierarchical stages.
}
\details{
We consider the \eqn{S}-stage hierarchical testing protocol outlined in Kim et al. (2007). Under this protocol, \eqn{N} individual specimens are first assigned to \eqn{m} non-overlapping pools, where each initial pool size is \eqn{c}; i.e., \eqn{N=mc}. The initial pools are tested in stage 1. If a pooled test is negative, all members in the pool are diagnosed as negative. However, if a pooled test is positive, the pool members are split into non-overlapping subpools to be tested in the next stage. This procedure is continued, while individual testing is typically used in the final stage, \eqn{S}, for case identification purposes. Note that when \eqn{N} is not divisible by \eqn{c}, a smaller remainder pool is used.

\code{S} is a positive integer, \code{S} >= 1. When \code{S}=1, only the non-overlapping initial pools are tested in stage 1.

If \code{N} is not divisible by the initial pool size \eqn{c}, we implement the following policy to test the remainder individuals: (1) If \code{S}=1, simply test the remainder pool once as a pooled sample; (2) If \code{S}>1, test the remainder pool based on 2-stage hierarchical testing using the last two values of \code{Se} and \code{Sp}.

\code{p} is a vector of individual disease probabilities. When all individuals have the same probability of disease, say, 0.10, p can be specified as p=rep(0.10, N) or p=0.10.

\code{psz} is a vector of length \code{S}, where the first element is the stage-1 pool size, the second element is the stage-2 pool size, and so on. Pool size at any stage must be divisible by the pool size used at the next stage. For example, \code{psz} can be specified as \code{c(12,3,1)} but not as \code{c(12,5,1)}.

When \code{S=1} (i.e., psz has length 1), only stage-1 testing is performed.

\code{Se} is a vector of length \code{S}, where the first element is the sensitivity of the assay used in stage 1, the second element is sensitivity of the assay in stage 2, and so on.

\code{Sp} is a vector of length \code{S}, where the first element is the specificity of the assay used in stage 1, the second element is specificity of the assay in stage 2, and so on.

\code{assayID} is a vector of length \code{S}, where the first element is the ID of the assay in stage 1, the second element is the ID of the assay in stage 2, and so on.

When available, the individual true disease statuses (1 for positive and 0 for negative) can be used in simulating the group testing data through argument \code{Yt}. When an input is entered for \code{Yt}, argument \code{p} will be ignored.
}
\examples{

library(groupTesting)

## Example 1: Two-stage hierarchical (Dorfman) testing
set.seed(123)
N <- 50              # Sample size
psz <- c(5, 1)       # Pool sizes used in stages 1 and 2
S <- 2               # The number of stages
Se <- c(0.95, 0.95)  # Sensitivities in stages 1-2
Sp <- c(0.98, 0.98)  # Specificities in stages 1-2
assayID <- c(1, 1)   # The same assay in both stages

# (a) Homogeneous population
pHom <- 0.10         # Overall prevalence
hier.gt.simulation(N=N,p=pHom,S=S,psz=psz,Se=Se,Sp=Sp,assayID=assayID)

# Alternatively, the individual true statuses can be used as: 
yt <- rbinom( N, size=1, prob=0.1 )
hier.gt.simulation(N=N,S=S,psz=psz,Se=Se,Sp=Sp,assayID=assayID,Yt=yt)

# (b) Heterogeneous population (regression)
param <- c(-3,2,1)
x1 <- rnorm(N, mean=0, sd=.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
pReg <- exp(X\%*\%param)/(1+exp(X\%*\%param)) # Inverse logit
hier.gt.simulation(N=N,p=pReg,S=S,psz=psz,Se=Se,Sp=Sp,assayID=assayID)

## Example 2: Initial (1-stage) pooled testing data
set.seed(123)
N <- 50
S <- 1
Se <- 0.95
Sp <- 0.98
assayID <- 1

# (a) Homogeneous population 
pHom <- 0.10

# a(i) Pooled testing
psz <- 5      
hier.gt.simulation(N,pHom,S,psz,Se,Sp,assayID)

# a(ii) Individual testing
psz <- 1       
hier.gt.simulation(N,pHom,S,psz,Se,Sp,assayID)

# (b) Heterogeneous population (regression)
param <- c(-3,2,1)
x1 <- rnorm(N, mean=0, sd=.75)
x2 <- rbinom(N, size=1, prob=0.5)
X <- cbind(1, x1, x2)
pReg <- stats::plogis(drop(X \%*\% param))

# b(i) Pooled testing
psz <- 5
hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)

# b(ii) Individual testing
psz <- 1
hier.gt.simulation(N,pReg,S,psz,Se,Sp,assayID)

## Example 3: Data with other configurations
set.seed(123)
N <- 48
p <- 0.10

# Initial pooled testing
pszH1 <- 4
hier.gt.simulation(N=N, p=p, S=1, psz=pszH1,
                   Se=.90, Sp=.96, assayID=1)

# Two-stage hierarchical testing
pszH2 <- c(4,1)
hier.gt.simulation(N=N, p=p, S=2, psz=pszH2,
                   Se=c(.90,.95), Sp=c(.96,.96), 
                   assayID=c(1,2))

# Four-stage hierarchical testing
pszH4 <- c(16,8,2,1)
hier.gt.simulation(N=N, p=p, S=4, psz=pszH4,
                   Se=c(.90,.95,.92,.90),
                   Sp=c(.96,.96,.90,.92),
                   assayID=1:4)

# Three-stage testing with assay IDs not in sequential order
pszH3 <- c(12,2,1)
hier.gt.simulation(N=N, p=p, S=3, psz=pszH3,
                   Se=c(.90,.95,.92),
                   Sp=c(.96,.96,.90),
                   assayID=c(2,1,3))

# Works with a remainder pool
N <- 50
psz <- c(12,2,1)
hier.gt.simulation(N=N, p=p, S=3, psz=psz,
                   Se=c(.90,.95,.92),
                   Sp=c(.96,.96,.90),
                   assayID=c(1,2,3))


}
\references{
Kim HY, Hudgens M, Dreyfuss J, Westreich D, and Pilcher C (2007). Comparison of Group Testing Algorithms for Case Identification in the Presence of Testing Error. \emph{Biometrics}, 63(4), 1152–1163.
}
\seealso{
\code{\link{array.gt.simulation}} for simulation of the array-based group testing data.
}
