% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkfns.R
\name{glmLink}
\alias{glmLink}
\title{Link Functions in the Class of Generalized Linear Models}
\usage{
glmLink(fn.name = c("logit", "probit", "cloglog"))
}
\arguments{
\item{fn.name}{One of the three: "logit", "probit", and "cloglog".}
}
\value{
A list with components:
\item{g}{The link function corresponding to "logit", "probit", or "cloglog".}
\item{dg}{The first derivative of g.}
\item{d2g}{The second derivative of g.}
\item{gInv}{The inverse of g.}
}
\description{
This function provides characteristics of common link functions (logit, probit, and comlementary log-log). Specifically, based on the link name, the function with its inverse, first derivative, and second derivative is provided.
}
\examples{

library(groupTesting)

## Try:
glmLink("logit")

}
